+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Fixed
 - Blendshapes naming in FBX so that multiple blendshapes all import correctly in Maya.
 - Don't override transforms when creating FBX Linked Prefab, so that the prefab updates properly when the FBX transforms are modified.
+- Changed FBX Linked Prefab to keep Unity materials instead of using materials exported to FBX file.
 
 ## [3.0.1-preview.2] - 2020-01-22
 ### Added